wip: add identicon and rename component for consistency
