+++ package.json
@@ -12,6 +12,7 @@
     "@material-icons/svg": "^1.0.28",
     "@polkadot/api": "9.11.1",
     "@polkadot/extension-dapp": "0.44.1",
+    "@polkadot/react-identicon": "^2.11.1",
     "@polkadot/ui-keyring": "^2.9.7",
     "@reach/tooltip": "^0.16.0",
     "@react-aria/accordion": "^3.0.0-alpha.14",
