+++ src/components/AccountSelect/index.tsx
@@ -1,0 +1,1 @@
+export { AccountSelector } from './AccountSelect';
