--- src/components/AccountSelector/index.tsx
@@ -1,1 +1,0 @@
-export { AccountSelector } from './AccountSelector';
